requirements.txt
awspricing==2.0.0
boto3==1.9.157
slackclient==2.0.1
